+++ Documentation/changelog.md
@@ -18,6 +18,7 @@
 * Add LCD menu for switch Sound [on - silent - off]
 * Add pause before deploy/stow for user confirmation
 * Add second serial for arduino due
+* Add PCF8574 Expansion IO for pin 120 - 121 - 122 - 123 - 124 - 125 - 126 - 127
 * Fix and clear code
 
 ### Version 4.3.6